SP-4997 Change "latest" to "master" or "main"
We still allow "latest", but mark them deprecated to encourage people
to explicitly say "main" or "master", because that gives a much better
idea of what they are actually getting.

This also plays more nicely with Spack's logic for versioning
constraints - e.g. "6.3:" includes "master", but doesn't include
"latest". This means that doing it this way requires fewer special
cases.

Also consistently set no_cache for such branch versions so that we
actually get "latest" versions even in the presence of caches.